Split KB creation wizard
